+++ app/src/app/globals.css
@@ -1,12 +1,2 @@
 @import "tailwindcss";
 @plugin "@tailwindcss/typography";
-@custom-variant dark (&:where(.dark, .dark *));
-
-.prose .html-table-output table th,
-.prose .html-table-output table td {
-  @apply text-center px-2 py-1 text-xs;
-}
-
-.prose .html-table-output p {
-  @apply text-right;
-}
