--- app/src/components/Icons.tsx
@@ -1,44 +1,0 @@
-export const CodePenIcon = ({ className = "" }) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    viewBox="0 0 24 24"
-    fill="currentColor" // Keep fill if you want it to inherit text color
-    // Use passed className for size/margin. Default internal classes don't need vertical-align.
-    className={`inline-block ${className}`}
-  >
-    {/* Path data... */}
-    <path
-      fillRule="evenodd"
-      clipRule="evenodd"
-      d="M11.605.862a.75.75 0 0 1 .79 0l10.5 6.5A.75.75 0 0 1 23.25 8v7a.75.75 0 0 1-.314.61l-10.5 7.5a.75.75 0 0 1-.872 0l-10.5-7.5A.75.75 0 0 1 .75 15V8a.75.75 0 0 1 .355-.638l10.5-6.5ZM2.25 13.654V9.457l3.147 2.248-3.147 1.949Zm9 7.389v-5.157L6.72 12.65l-3.867 2.395 8.397 5.998Zm1.5 0v-5.157l4.53-3.236 3.867 2.395-8.397 5.998Zm-4.647-9.248L12 14.578l3.897-2.783L12 9.382l-3.897 2.413Zm10.5-.09 3.147 1.949V9.457l-3.147 2.248Zm-1.383-.855-4.47-2.768V2.846l8.397 5.199-3.927 2.805Zm-10.44 0 4.47-2.768V2.846L2.853 8.045l3.927 2.805Z"
-    />
-  </svg>
-);
-
-export const ColabIcon = ({ className = "" }) => (
-  <svg
-    xmlns="http://www.w3.org/2000/svg"
-    viewBox="0 0 48 48"
-    // Use passed className for size/margin. Default internal classes don't need vertical-align.
-    className={`inline-block ${className}`}
-    // Removed fill="currentColor" if not desired, Colab has specific fills
-  >
-    {/* Path data... */}
-    <path
-      fill="#ffb300"
-      d="M33.5,10C26.044,10,20,16.044,20,23.5C20,30.956,26.044,37,33.5,37S47,30.956,47,23.5 C47,16.044,40.956,10,33.5,10z M33.5,30c-3.59,0-6.5-2.91-6.5-6.5s2.91-6.5,6.5-6.5s6.5,2.91,6.5,6.5S37.09,30,33.5,30z"
-    />
-    <path
-      fill="#ffb300"
-      d="M19.14,28.051l0-0.003C17.96,29.252,16.318,30,14.5,30C10.91,30,8,27.09,8,23.5s2.91-6.5,6.5-6.5  c1.83,0,3.481,0.759,4.662,1.976l3.75-6.024C20.604,11.109,17.683,10,14.5,10C7.044,10,1,16.044,1,23.5C1,30.956,7.044,37,14.5,37 c3.164,0,6.067-1.097,8.369-2.919L19.14,28.051z"
-    />
-    <path
-      fill="#f57c00"
-      d="M8,23.5c0-1.787,0.722-3.405,1.889-4.58l-4.855-5.038C2.546,16.33,1,19.733,1,23.5  c0,3.749,1.53,7.14,3.998,9.586l4.934-4.964C8.74,26.944,8,25.309,8,23.5z"
-    />
-    <path
-      fill="#f57c00"
-      d="M38.13,18.941C39.285,20.114,40,21.723,40,23.5c0,3.59-2.91,6.5-6.5,6.5  c-1.826,0-3.474-0.755-4.655-1.968l-4.999,4.895C26.298,35.437,29.714,37,33.5,37C40.956,37,47,30.956,47,23.5  c0-3.684-1.479-7.019-3.871-9.455L38.13,18.941z"
-    />
-  </svg>
-);
